+++ install.sh
@@ -106,7 +106,7 @@
   REPO="https://github.com/manifoldco/manifold-cli"
 
   if [ -z "$MANIFOLD_VERSION" ]; then
-    MANIFOLD_VERSION="0.9.3"
+    MANIFOLD_VERSION="0.10.0"
   fi
 
   if [ -z "$MANIFOLD_DIR" ]; then